+++ FWI/R/make_minmax.r
@@ -26,12 +26,19 @@ temp_min_max <- function(temp_noon, rh_noon) {
 #'
 #' @param   df          daily noon LST weather stream, columns:
 #'                      yr, mon, day, temp, rh, ws, prec
+#' @param   silent      suppresses informative print statements (default False)
 #' @param   round_out   decimals to truncate output to, NA for none (default 4)
 #' @return              daily min/max weather stream, columns:
 #'                      yr, mon, day, temp_min, temp_max, rh_min, rh_max,
 #'                      ws_min, ws_max, prec
 #' @export  daily_to_minmax
-daily_to_minmax <- function(df, round_out = 4) {
+daily_to_minmax <- function(df, silent = FALSE, round_out = 4) {
+  if (!silent) {
+    writeLines("\n########")
+    writeLines(paste0("FWI2025: Make Min/Max Inputs (", version(), ")\n"))
+    writeLines("Predicting daily min/max weather")
+  }
+
   # check df_wx class for data.frame or data.table
   wasDT <- is.data.table(df)
   if (wasDT) {
@@ -73,13 +80,17 @@ daily_to_minmax <- function(df, round_out = 4) {
   if (!wasDT) {
     setDF(df)
   }
+
+  if (!silent) {
+    writeLines("########\n")
+  }
   return(df)
 }
 
 
 # run daily_to_minmax by command line via Rscript
-# required args: input csv and output csv
-# optional arg: round_out
+# required arguments: input csv, output csv
+# optional arguments: silent, round_out
 if ("--args" %in% commandArgs() && sys.nframe() == 0) {
   args <- commandArgs(trailingOnly = TRUE)
   if (length(args) < 2) {
@@ -87,11 +98,13 @@ if ("--args" %in% commandArgs() && sys.nframe() == 0) {
   }
   input <- args[1]
   output <- args[2]
-  if (length(args) >= 3) round_out <- args[3]
+  if (length(args) >= 3) silent <- as.logical(args[3])
+  else silent <- FALSE
+  if (length(args) >= 4) round_out <- args[4]
   else round_out <- 4
-  if (length(args) >= 4) warning("Too many input arguments provided, some unused")
+  if (length(args) >= 5) warning("Too many input arguments provided, some unused")
 
   df_in <- read.csv(input)
-  df_out <- daily_to_minmax(df_in, round_out)
+  df_out <- daily_to_minmax(df_in, silent, round_out)
   write.csv(df_out, output, row.names = FALSE)
 }